### 2018-09-23 #972 @whedon
[PRE REVIEW]: Lithology: A Landlab submodule for spatially variable rock properties
**Submitting author:** @kbarnhart (<a href="http://orcid.org/0000-0001-5682-455X">Katherine Barnhart</a>)
**Repository:** <a href="https://github.com/landlab/landlab" target ="_blank">https://github.com/landlab/landlab</a>
**Version:** 1.5.4
