### 2018-09-23 #972 @arfon
> - I wrote version 1.5.4 because this is the number of the first release version of Landlab that I anticipate this submission being a part of (our next release is in December, we are currently at v1.5.3, and this submission is on the master branch). I haven't tagged a commit because that triggers automatic building of packages.
> - This submission only pertains to the Lithology submodule of Landlab (located at landlab/landlab/components/lithology), not Landlab as a whole. The paper.md is located at landlab/landlab/components/lithology/joss/paper.md. I discussed the appropriate way to handle this submission at Issue #433 and in the interim of implementing a more formal method for specifying the submodule of interest, @arfon recommended that I just make a note here.

Notes from the submitting author ☝️
