### 2018-09-23 #972 @kbarnhart
A couple of questions

1. I've noticed that there is an error in the proofs associated with my not understanding how to the citations correctly.

How do I cite multiple people at the same time, or add e.g. to the beginning of the citation?

I had wrote:
`` (e.g. [@Forte2016Complexites], [@Perne2017Steady])``

Should I have done?
``[e.g., @Forte2016Complexites], [@Perne2017Steady])``

or something else?

2. Should I set up a reserve Zenodo DOI at this point, or is OK to wait to do this until later in the review process?

3. To fix these things do I just make changes on the the Landlab `master` branch and ask `whedon` to recompile the pdf?

thanks!
