### 2018-09-24 #972 @kthyng
Hi @kbarnhart!

> Should I set up a reserve Zenodo DOI at this point, or is OK to wait to do this until later in the review process?

You can do this at the end of the review process.

> To fix these things do I just make changes on the the Landlab master branch and ask whedon to recompile the pdf?

Yes I think so to the master branch part. You can make your changes now and then when your Review issue is opened, `whedon` will try to compile your paper at that time again. I'm not sure you have the authority to ask `whedon` to recompile, but I can do it anyway (or at least I can try!).

Do you have any suggestions for reviewers? 2 to 4 would be great, so we can get 2 committed reviewers. There is a [list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit?usp=sharing) of potential reviewers available you can choose from if you don't have anyone in mind.
