### 2018-09-24 753e946
remove proj from setup.py
