### 2018-09-25 #972 @sgrieve
I'm afraid I wont have time to review this one, since I'm supposed to be working on another JOSS review at the moment. Another potential reviewer might be @dvalters
