### 2018-09-25 0fde578
debug tests
