### 2018-09-25 0a90ba6
Restore flow direction implementation test
