+++ landlab/components/stream_power/sed_flux_dep_incision.py
@@ -392,6 +392,15 @@ class SedDepEroder(Component):
                  # params for model numeric behavior:
                  pseudoimplicit_repeats=50, **kwds):
         """Constructor for the class."""
+        if 'flow__receiver_node' in grid.at_node:
+            if (grid.at_node['flow__receiver_node'].size != grid.size('node')):
+                msg = ('A route-to-multiple flow director has been '
+                       'run on this grid. The landlab development team has not '
+                       'verified that FastscapeEroder is compatible with '
+                       'route-to-multiple methods. Please open a GitHub Issue '
+                       'to start this process.')
+                raise NotImplementedError(msg)
+
         self._grid = grid
         self.pseudoimplicit_repeats = pseudoimplicit_repeats
 
