### 2018-09-25 89a96b6
add basic docs to ErosionDeposition
These docs should provide a general user with enough information to get
the component to run and vaguely follow what it’s doing.
