### 2018-09-25 d24da51
PEP8
Quick scan
