### 2018-09-25 #972 @kthyng
Hi @speleophysics and @tristan-salles! Are you willing and able to review this submission to JOSS? You can read about the review process [here](https://joss.readthedocs.io/en/latest/review_criteria.html).
