### 2018-09-25 #972 @speleophysics
I can review it, though I do have two other reviews on my plate at the
moment. What is the suggested timeline?
-------------------------
Matt Covington
Associate Professor
Department of Geosciences
340 N. Campus Dr.
216 Gearhart Hall
University of Arkansas
Fayetteville, AR 72701

webpage - http://www.speleophysics.com/
-------------------------

On Tue, Sep 25, 2018 at 11:22 AM Kristen Thyng <notifications@github.com>
wrote:

> Hi @speleophysics <https://github.com/speleophysics> and @tristan-salles
> <https://github.com/tristan-salles>! Are you willing and able to review
> this submission to JOSS? You can read about the review process here
> <https://joss.readthedocs.io/en/latest/review_criteria.html>.
>
