### 2018-09-25 484952e
Update readme
Addresses Issue #798
