### 2018-09-26 landlab/landlab#798 @kbarnhart
Lithology JOSS proof citations are not formatted correctly
The [PDF proof of the Lithology manuscript](https://github.com/openjournals/joss-papers/blob/joss.00979/joss.00979/10.21105.joss.00979.pdf) has citations that are not formatted correctly.

This Issue is associated with [JOSS Review # 979](https://github.com/openjournals/joss-reviews/issues/979)
