### 2018-09-26 9eff10d
initialise new component
Run method still absent
