+++ landlab/components/stream_power/sediment_transport_stream_power.py
@@ -1,0 +1,160 @@
+import numpy as np
+from landlab.components.erosion_deposition.generalized_erosion_deposition import (_GeneralizedErosionDeposition,
+                                                            DEFAULT_MINIMUM_TIME_STEP)
+from landlab.utils.return_array import return_array_at_node
+from .cfuncs import calculate_qs_in
+
+ROOT2 = np.sqrt(2.0)    # syntactic sugar for precalculated square root of 2
+TIME_STEP_FACTOR = 0.5  # factor used in simple subdivision solver
+DEFAULT_MINIMUM_TIME_STEP = 0.001  # default minimum time step duration
+
+
+class TransportLimitedEroder(_GeneralizedErosionDeposition):
+    """
+    Implementation of a river erosion component which treats fluvial sediment
+    flux as a function of stream power, and then uses mass balance (i.e.,
+    the divergence of sediment flux) to dictate the response of the river bed.
+
+    In other words,
+
+    Qs = K * A**m * S**n - optional_threshold
+    dz/dt = d/dx(Qs)
+
+    If provided, the threshold is implemented smoothly. See the
+    StreamPowerSmoothThresholdEroder for more details.
+    """
+    _name = 'TransportLimitedEroder'
+
+    _input_var_names = (
+        'flow__receiver_node',
+        'flow__upstream_node_order',
+        'topographic__steepest_slope',
+        'drainage_area',
+    )
+
+    _output_var_names = (
+        'topographic__elevation'
+    )
+
+    _var_units = {
+        'flow__receiver_node': '-',
+        'flow__upstream_node_order': '-',
+        'topographic__steepest_slope': '-',
+        'drainage_area': 'm**2',
+        'topographic__elevation': 'm',
+    }
+
+    _var_mapping = {
+        'flow__receiver_node': 'node',
+        'flow__upstream_node_order': 'node',
+        'topographic__steepest_slope': 'node',
+        'drainage_area': 'node',
+        'topographic__elevation': 'node',
+    }
+
+    _var_doc = {
+        'flow__receiver_node':
+            'Node array of receivers (node that receives flow from current '
+            'node)',
+        'flow__upstream_node_order':
+            'Node array containing downstream-to-upstream ordered list of '
+            'node IDs',
+        'topographic__steepest_slope':
+            'Topographic slope at each node',
+        'drainage_area':
+            "Upstream accumulated surface area contributing to the node's "
+            "discharge",
+        'topographic__elevation':
+            'Land surface topographic elevation',
+    }
+
+    def __init__(self, grid, K=None, phi=None, v_s=None,
+                 m_sp=None, n_sp=None, sp_crit=0.0, F_f=0.0,
+                 discharge_field=None,
+                 solver='adaptive',
+                 dt_min=DEFAULT_MINIMUM_TIME_STEP,
+                 **kwds):
+        """Initialize the TransportLimitedEroder component.
+
+        Parameters
+        ----------
+        grid : ModelGrid
+            Landlab ModelGrid object
+        K : float, field name, or array
+            Prefactor for sediment transport equation (units vary).
+        phi : float
+            Sediment porosity [-].
+        m_sp : float
+            Drainage area exponent for sediment transport equation [-]
+        n_sp : float
+            Slope exponent for sediment transport equation [-]
+        sp_crit : float, field name, or array
+            Critical stream power to erode substrate [E/(TL^2)]
+        F_f : float
+            Fraction of eroded material that turns into "fines" that do not
+            contribute to (coarse) sediment load. Defaults to zero.
+        discharge_field : float, field name, or array, or None
+            If None, the component will use its standard equation for sediment
[…]
