+++ landlab/grid/divergence.py
@@ -4,7 +4,6 @@ import numpy as np
 from landlab.utils.decorators import use_field_name_or_array
 
 
-
 @use_field_name_or_array('link')
 def calc_flux_div_at_node(grid, unit_flux, out=None):
     """Calculate divergence of link-based fluxes at nodes.
@@ -51,8 +50,8 @@ def calc_flux_div_at_node(grid, unit_flux, out=None):
     >>> calc_flux_div_at_node(rg, 'neg_grad_at_link')
     array([ 0.  ,  0.  ,  0.  ,  0.  ,  0.  ,  1.64,  0.94,  0.  ,  0.  ,
             0.  ,  0.  ,  0.  ])
-            
-            
+
+
     Notes
     -----
     Performs a numerical flux divergence operation on nodes.
@@ -67,9 +66,8 @@ def calc_flux_div_at_node(grid, unit_flux, out=None):
     elif out.size != grid.number_of_nodes:
         raise ValueError('output buffer length mismatch with number of nodes')
 
-    out[grid.node_at_cell] = _calc_net_face_flux_at_cell(grid, 
-                                    unit_flux[grid.link_at_face]) \
-                                    / grid.area_of_cell
+    out[grid.node_at_cell] = _calc_net_face_flux_at_cell(
+        grid, unit_flux[grid.link_at_face]) / grid.area_of_cell
 
     return out
 
@@ -78,10 +76,10 @@ def calc_flux_div_at_node(grid, unit_flux, out=None):
 def calc_flux_div_at_cell(grid, unit_flux, out=None):
     """Calculate divergence of link-based fluxes at cells.
 
-    This function is very similar to the function *calc_flux_div_at_node*. 
-    
-    Given a flux per unit width across each cell face in the grid, calculate 
-    the net outflux (or influx, if negative) divided by cell area, at each 
+    This function is very similar to the function *calc_flux_div_at_node*.
+
+    Given a flux per unit width across each cell face in the grid, calculate
+    the net outflux (or influx, if negative) divided by cell area, at each
     cell.
 
     Parameters
@@ -124,7 +122,7 @@ def calc_flux_div_at_cell(grid, unit_flux, out=None):
     >>> _ = rg.add_field('neg_grad_at_link', -lg, at = 'link')
     >>> calc_flux_div_at_cell(rg, 'neg_grad_at_link')
     array([ 1.64,  0.94])
-            
+
     Notes
     -----
     Performs a numerical flux divergence operation at cells.
@@ -140,12 +138,11 @@ def calc_flux_div_at_cell(grid, unit_flux, out=None):
         raise ValueError('output buffer length mismatch with number of cells')
 
     if unit_flux.size == grid.number_of_links:
-        out = _calc_net_face_flux_at_cell(grid, 
-                                    unit_flux[grid.link_at_face]) \
-                                    / grid.area_of_cell
+        out = _calc_net_face_flux_at_cell(
+            grid, unit_flux[grid.link_at_face]) / grid.area_of_cell
     elif unit_flux.size == grid.number_of_faces:
-        out = _calc_net_face_flux_at_cell(grid, unit_flux) \
-                                / grid.area_of_cell
+        out = _calc_net_face_flux_at_cell(
+            grid, unit_flux) / grid.area_of_cell
 
     return out
 
@@ -224,8 +221,8 @@ def calc_net_flux_at_node(grid, unit_flux_at_links, out=None):
     if out is None:
         out = grid.zeros(at='node')
 
-    out[grid.node_at_cell] = _calc_net_face_flux_at_cell(grid, 
-                                unit_flux_at_links[grid.link_at_face])
+    out[grid.node_at_cell] = _calc_net_face_flux_at_cell(
+        grid, unit_flux_at_links[grid.link_at_face])
     return out
 
[…]
