### 2018-09-27 d88a017
A lot closer to complete
This *may* be a close to final version
