+++ landlab/components/stream_power/sediment_transport_stream_power.py
@@ -22,6 +22,12 @@ class TransportLimitedEroder(_GeneralizedErosionDeposition):
 
     If provided, the threshold is implemented smoothly. See the
     StreamPowerSmoothThresholdEroder for more details.
+
+    The component permits supply of a porosity (phi) and wash fraction (F_f).
+    Note that if these are not 0., then you are implicitly assuming that
+    any sediment pickup is occurring out of intact rock, that has not
+    previously been processed by a river. This assumption may often not be
+    appropriate.
     """
     _name = 'TransportLimitedEroder'
 
@@ -124,17 +130,24 @@ class TransportLimitedEroder(_GeneralizedErosionDeposition):
                 dt_min=dt_min, discharge_field='drainage_area')
 
         self._grid = grid  # store grid
+        self._one_by_erosion_loss = 1. / (F_f * phi)
+        # the net loss of rock during erosion
 
         # K's and critical values can be floats, grid fields, or arrays
         self.K = return_array_at_node(grid, K)
+        self.Qs_in = grid.zeros('node', dtype=float)
         # special cases if sp_crit is 0 or <0...
         self.sp_crit = return_array_at_node(grid, sp_crit)
         if np.any(self.sp_crit < 0.):
             raise ValueError('sp_crit must be >= 0. everywhere')
         if np.allclose(sp_crit, 0.):
-            self._calc_trp = self._calc_transport_rates_wo_thresh()
+            self._calc_trp = self._calc_transport_rates_wo_thresh
         else:
-            self._calc_trp = self._calc_transport_rates_with_thresh()
+            self._calc_trp = self._calc_transport_rates_with_thresh
+        if np.isclose(self._one_by_erosion_loss, 1.):
+            self._calc_sed_div = _calc_sed_flux_divergence
+        else:
+            self._calc_sed_div = _calc_sed_flux_divergence_lossy
 
         # Handle option for solver
         if solver == 'basic':
@@ -158,3 +171,151 @@ class TransportLimitedEroder(_GeneralizedErosionDeposition):
     def _calc_transport_rates_wo_thresh(self):
         """Calculate Qs with a threshold"""
         self.qs = self.K * self.Q_to_the_m * np.power(self.slope, self.n_sp)
+
+    def _calc_discharges(self, flooded_nodes):
+        """
+        Calculate sediment discharge out of a node in channel.
+        """
+        self._calc_trp()  # this sets fluxes, so need to weight by A
+        self.Qs_out = self.qs * self.cell_area_at_node
+        if flooded_nodes is not None:
+            self.Qs_out[flooded_nodes] = 0.
+
+    def run_with_adaptive_time_step_solver(self, dt=1.0, flooded_nodes=[],
+                                           **kwds):
+        """CHILD-like solver that adjusts time steps to prevent slope
+        flattening."""
+
+        # Initialize remaining_time, which records how much of the global time
+        # step we have yet to use up.
+        remaining_time = dt
+
+        # z = self._grid.at_node['topographic__elevation']
+        r = self.flow_receivers
+        dzdt = self.grid.zeros('node', dtype=float)
+        cores = self.grid.core_nodes
+
+        first_iteration = True
+
+        # Outer WHILE loop: keep going until time is used up
+        while remaining_time > 0.0:
+
+            # Update all the flow-link slopes.
+            #
+            # For the first iteration, we assume this has already been done
+            # outside the component (e.g., by flow router), but we need to do
+            # it ourselves on subsequent iterations.
+            if not first_iteration:
+                # update the link slopes
+                self._update_flow_link_slopes()
+                # update where nodes are flooded. This shouuldn't happen bc
+                # of the dynamic timestepper, but just incase, we update here.
+                new_flooded_nodes = np.where(self.slope < 0)[0]
[…]
