+++ landlab/components/stream_power/tests/test_TL.py
@@ -5,11 +5,13 @@ from numpy.testing import assert_array_almost_equal
 
 from landlab import RasterModelGrid
 from landlab.components import TransportLimitedEroder
+from landlab.components.sediment_transport_stream_power import _calc_sed_flux_divergence
+from landlab.components.sediment_transport_stream_power import _calc_sed_flux_divergence_lossy
 from landlab.components import FlowAccumulator
 
 
 def test_multidirection():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     z = mg.add_zeros('node', 'topographic__elevation')
     fa = FlowAccumulator(mg, flow_director='D8')
     fa.run_one_step()
@@ -18,43 +20,43 @@ def test_multidirection():
 
 
 def test_bad_init1():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     with pytest.raises(ValueError):
         TL = TransportLimitedEroder(mg, phi=1.)
 
 
 def test_bad_init2():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     with pytest.raises(ValueError):
         TL = TransportLimitedEroder(mg, phi=-1.)
 
 
 def test_bad_init3():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     with pytest.raises(ValueError):
         TL = TransportLimitedEroder(mg, F_f=1.1)
 
 
 def test_bad_init4():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     with pytest.raises(ValueError):
         TL = TransportLimitedEroder(mg, F_f=-1.)
 
 
 def test_bad_init5():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     with pytest.raises(ValueError):
         TL = TransportLimitedEroder(mg, solver='nope')
 
 
 def test_bad_init6():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     with pytest.raises(ValueError):
         TL = TransportLimitedEroder(mg, sp_crit=-1.)
 
 
 def test_bad_init7():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     badcrits = np.zeros(16, dtype=float)
     badcrits[5] = -1.
     with pytest.raises(ValueError):
@@ -62,7 +64,7 @@ def test_bad_init7():
 
 
 def test_bad_init8():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     badcrits = np.zeros(16, dtype=float)
     badcrits[5] = -1.
     crits = mg.add_field('node', 'crits', badcrits)
@@ -71,15 +73,84 @@ def test_bad_init8():
 
 
 def test_sed_flux_in():
-    mg = RasterModelGrid(4, 4, 1)
+    mg = RasterModelGrid((4, 4), 1.)
     Qsin = mg.add_ones('node', 'sediment__flux')
     TL = TransportLimitedEroder(mg)
     assert np.allclose(TL.qs, 1.)
 
 
+def test_flux_transfer():
+    mg = RasterModelGrid((3, 3), 1.)
+    z = np.array([9., 8., 6.,
+                  7., 3., 2.,
+                  5., 0., 1.])
+    elev_order = np.array([0, 1, 3, 2, 6, 4, 5, 8, 7])
[…]
