### 2018-09-27 9dfc43a
Tests so far pass
As yet, no tests on the main method
