+++ landlab/components/stream_power/__init__.py
@@ -2,7 +2,8 @@ from .stream_power import StreamPowerEroder
 from .fastscape_stream_power import FastscapeEroder
 from .stream_power_smooth_threshold import StreamPowerSmoothThresholdEroder
 from .sed_flux_dep_incision import SedDepEroder
+from .sediment_transport_stream_power import TransportLimitedEroder
 
 
 __all__ = ['StreamPowerEroder', 'FastscapeEroder', 'SedDepEroder',
-           'StreamPowerSmoothThresholdEroder']
+           'StreamPowerSmoothThresholdEroder', 'TransportLimitedEroder']
