+++ landlab/components/stream_power/tests/test_TL.py
@@ -5,77 +5,105 @@ from numpy.testing import assert_array_almost_equal
 
 from landlab import RasterModelGrid
 from landlab.components import TransportLimitedEroder
-from landlab.components.sediment_transport_stream_power import _calc_sed_flux_divergence
-from landlab.components.sediment_transport_stream_power import _calc_sed_flux_divergence_lossy
+from landlab.components.stream_power.sediment_transport_stream_power import _calc_sed_flux_divergence
+from landlab.components.stream_power.sediment_transport_stream_power import _calc_sed_flux_divergence_lossy
 from landlab.components import FlowAccumulator
 
 
 def test_multidirection():
-    mg = RasterModelGrid((4, 4), 1.)
+    mg = RasterModelGrid((5, 5), 1.)
     z = mg.add_zeros('node', 'topographic__elevation')
-    fa = FlowAccumulator(mg, flow_director='D8')
+    z += np.random.rand(25)
+    fa = FlowAccumulator(mg, flow_director='MFD')
     fa.run_one_step()
     with pytest.raises(NotImplementedError):
-        TL = TransportLimitedEroder(mg)
+        TL = TransportLimitedEroder(mg, K=1.e-5)
 
 
 def test_bad_init1():
     mg = RasterModelGrid((4, 4), 1.)
+    z = mg.add_zeros('node', 'topographic__elevation')
+    fa = FlowAccumulator(mg, flow_director='D8')
+    fa.run_one_step()
     with pytest.raises(ValueError):
-        TL = TransportLimitedEroder(mg, phi=1.)
+        TL = TransportLimitedEroder(mg, K=1.e-5, phi=1.)
 
 
 def test_bad_init2():
     mg = RasterModelGrid((4, 4), 1.)
+    z = mg.add_zeros('node', 'topographic__elevation')
+    fa = FlowAccumulator(mg, flow_director='D8')
+    fa.run_one_step()
     with pytest.raises(ValueError):
-        TL = TransportLimitedEroder(mg, phi=-1.)
+        TL = TransportLimitedEroder(mg, K=1.e-5, phi=-1.)
 
 
 def test_bad_init3():
     mg = RasterModelGrid((4, 4), 1.)
+    z = mg.add_zeros('node', 'topographic__elevation')
+    fa = FlowAccumulator(mg, flow_director='D8')
+    fa.run_one_step()
     with pytest.raises(ValueError):
-        TL = TransportLimitedEroder(mg, F_f=1.1)
+        TL = TransportLimitedEroder(mg, K=1.e-5, F_f=1.1)
 
 
 def test_bad_init4():
     mg = RasterModelGrid((4, 4), 1.)
+    z = mg.add_zeros('node', 'topographic__elevation')
+    fa = FlowAccumulator(mg, flow_director='D8')
+    fa.run_one_step()
     with pytest.raises(ValueError):
-        TL = TransportLimitedEroder(mg, F_f=-1.)
+        TL = TransportLimitedEroder(mg, K=1.e-5, F_f=-1.)
 
 
 def test_bad_init5():
     mg = RasterModelGrid((4, 4), 1.)
+    z = mg.add_zeros('node', 'topographic__elevation')
+    fa = FlowAccumulator(mg, flow_director='D8')
+    fa.run_one_step()
     with pytest.raises(ValueError):
-        TL = TransportLimitedEroder(mg, solver='nope')
+        TL = TransportLimitedEroder(mg, K=1.e-5, solver='nope')
 
 
 def test_bad_init6():
     mg = RasterModelGrid((4, 4), 1.)
[…]
