### 2018-09-27 #979 @kbarnhart
@kthyng Looks like I have the authority to ask for pdf recompilation. It also looks like my PR fixed the citation formatting.
