### 2018-09-28 0608d2f
part of the issue comes from an equals sign
