### 2018-10-01 2833245
Interim tweaks
Minor debug
