### 2018-10-02 6d9f2ed
using a generic stopping basis.
