+++ landlab/components/profiler/ridge_profiler.py
@@ -10,7 +10,7 @@ from landlab.components.profiler.base_profiler import _NetworkProfiler
 class RidgeProfiler(_NetworkProfiler):
     """
     """
-    def __init__(self, grid, threshold=None):
+    def __init__(self, grid, stopping_field='drainage_area', threshold=None):
         """
         """
         super(RidgeProfiler, self).__init__(grid)
@@ -20,7 +20,7 @@ class RidgeProfiler(_NetworkProfiler):
             threshold = 4. * np.amin(grid.area_of_cell)
         self.threshold = threshold
 
-        self._ridge_mask = self._drainage_area < self.threshold
+        self._ridge_mask = self._stopping_field < self.threshold
 
 
     def something(self):
