### 2018-10-03 #979 @tristan-salles
I am pleased to recommend publication of this additional new module to Landlab!

This is clearly going to be of interest to many of the existing users of this code and has also the potential to attract more researchers who are interested in adding more realism in their models with spatially variable lithologies producing variable erodability and diffusivity.

Following are some of my comments and minor issues that I've found during the review process.

### Version number

The current Landlab Github release is 1.5.3 and therefore does not match the proposed release 1.5.4. That is the reason why I didn't click it the checklist. Yet it looks like this is going to be handle later and the relevant version will be release after the initial JOSS submission.

### Functionality & documentation

There isn't _per se_ a specific installation for this submodule but there is already existing and well-establish documentation regarding the installation of Landlab. I did pull the specified request number [#674](Pull Request # 674) from  Landlab repository and didn't have any issue with the local installation following the `pip` method and everything worked as expected.

I have downloaded the corresponding Jupyter notebooks (linked provided in the manuscript). I followed the instructions regarding the installation of additional packages:
+ xarray
+ HoloViews
+ Bokeh

and was able to run all the cells without issues.

The tutorial and step-by-step guide to illustrate the capability of this new submodule are well explained and illustrative as well as simple. I've noticed that the Jupyter notebook is quite large >(>60 MB) maybe it will be better to clear all cells before uploading the notebook to its repo on Github but you might want to keep the plots... Also the two last links to the readme website for the 2 lithology objects are not working and should be updated.

As already pointed above the submodule is part of the Landlab software and contains the requirement from the review checklist:
+ statement of need
+ clearly-stated list of dependencies
+ the Jupyter notebooks from the manuscript present an example usage
+ an API documentation for this particular submodule is provided and properly referenced in the manuscript.
+ automated tests are part of Landlab and I assume are ran for this submodule as well
+ there is clear community guidelines statement

### Additional minor comments:

+ in the paper in addition to pointing to the _Landlab_ GitHub main repository I will also suggest to add a link to the _Lithology_ [repo](https://github.com/landlab/landlab/tree/master/landlab/components/lithology)
+ in the `lithology` class the documentation has the word 'corresponding' misspelled (cooresponding)  several times. There is also a space that could be deleted just after 'number-of' in the 'number-of -nodes' in the following sentence.
+ in the ``run_one_step`` method I've noticed a typo on the first parameter name called `dz_advenction` instead of `dz_advection`.
