### 2018-10-03 landlab/landlab#807 @kbarnhart
Misc. raised in JOSS review
This issue describes miscellaneous issues outlined in the [JOSS review for the Lithology Component](https://github.com/openjournals/joss-reviews/issues/979).
