### 2018-10-03 landlab/landlab#807 @kbarnhart
Minor comments from @tristan-salles

* in the paper in addition to pointing to the Landlab GitHub main repository I will also suggest to add a link to the Lithology [repo](https://github.com/landlab/landlab/tree/master/landlab/components/lithology)
* in the `lithology` class the documentation has the word 'corresponding' misspelled (cooresponding) several times. There is also a space that could be deleted just after 'number-of' in the 'number-of -nodes' in the following sentence.
* in the `run_one_step` method I've noticed a typo on the first parameter name called `dz_advenction` instead of `dz_advection`.
