### 2018-10-03 landlab/landlab#807 @kbarnhart
Additional comments from @Tristan-salles to be addressed:

* I've noticed that the Jupyter notebook is quite large >(>60 MB) maybe it will be better to clear all cells before uploading the notebook to its repo on Github but you might want to keep the plots...

* Also the two last links to the readme website for the 2 lithology objects are not working and should be updated.
