### 2018-10-03 landlab/landlab#807 @kbarnhart
@gregtucker @mcflugen @nicgaspar do you have thoughts on clearing the cells in this notebook. It would mean that it would be much smaller, but it would also mean that a user couldn't preview the notebook on the web [like you can now](https://nbviewer.jupyter.org/github/landlab/tutorials/blob/next/lithology/lithology_and_litholayers.ipynb).

It is 60 MB at present.
