+++ landlab/components/flow_accum/tests/test_flow_accumulator.py
@@ -1025,5 +1025,18 @@ def test_hex_mfd():
     fa = FlowAccumulator(mg, flow_director="MFD")
     fa.run_one_step()
 
-def test_flat_grids():
-    pass
+
+def test_flat_grids_all_directors():
+    for fd in ['FlowDirectorSteepest',
+               'FlowDirectorD8',
+               'FlowDirectorDINF',
+               'FlowDirectorMFD']:
+        mg = RasterModelGrid(10, 20, dx=1)
+        z = mg.add_zeros('topographic__elevation', at='node')
+        fa = FlowAccumulator(mg, flow_director=fd)
+        try:
+            fa.run_one_step()
+            print(fd + ' success')
+        except IndexError:
+            print(fd + ' failed')
+        del mg, z, fa
