### 2018-10-04 landlab/landlab#807 @nicgaspar
60 mb is a bit big, but not too obnoxious (at least to me). I'd be for keeping in the results, as I've now been convinced that having them there to see without needing to run the notebook is useful.

Thanks @kbarnhart
