### 2018-10-04 #979 @speleophysics
I am happy to recommend this new Landlab module for publication. In general, I agree with @tristan-salles assessment and comments. The new module is a nice contribution that will undoubtedly be used widely within the community. The purpose is expressed clearly in the accompanying paper, and the code works as advertised.

I have read through the documentation and gone through the tutorials. In the process, I found a few erroneous links to the documentation in the tutorials and have made a few suggestions for clarification (provided as issues in the landlab/tutorials repo, https://github.com/landlab/tutorials/issues/38#issue-366912619,  https://github.com/landlab/tutorials/issues/36#issue-366512363). However, overall, the documentation is well-written and easy to understand.

Regarding the checklist above, there are two items I have not yet checked: 1) as noted by @tristan-salles , the version number is not yet updated, 2) A few of the references in the paper are missing DOIs.
