### 2018-10-05 #979 @kbarnhart
Thanks to both @tristan-salles and @speleophysics for constructive and thoughtful reviews. I anticipate finishing revisions by early next week. I'll make a note here when I'm done.
