### 2018-10-08 132ede3
add link to lithology repository
