+++ landlab/components/lithology/joss/paper.md
@@ -39,7 +39,7 @@ Landlab is an Open Source Python package that provides a framework for the devel
 
 This contribution describes ``Lithology``, a Landlab submodule designed to support the representation of 3D variations in rock material properties within the Landlab framework. It includes two classes: ``Lithology`` is a generic representation of spatially varying rock material, and ``LithoLayers`` is a derived class that treats parallel layers of material variations. In both classes, each rock type may have multiple attributes. Rock layers may be removed through erosion, or added to through deposition. Two options for the underlying datastructure are supported: "event layers", in which the data structure stores each time-step as an event even if there is no material in the layer, or "material layers", in which entries in the datastructure represent contiguous material of the same property, but not necessarily the same age. This second option is more memory efficient but does not record the transient dynamics of erosion and deposition.
 
-Source code for ``Lithology`` and ``Litholayers`` is available as part of the [Landlab python package](https://github.com/landlab/landlab). The ``Lithology`` submodule is documented using Docstrings, and the documentation can be found on the Landlab ReadTheDocs site. One page exists for the [Lithology component](https://landlab.readthedocs.io/en/master/landlab.components.lithology.html) and a second for the [LithoLayers component](https://landlab.readthedocs.io/en/master/landlab.components.litholayers.html). Unit and docstring tests provide 100% coverage of this submodule. [Pull Request # 674](https://github.com/landlab/landlab/pull/674) brought the ``Lithology`` submodule into the core Landlab source code. A tagged version of Landlab that includes the ``Lithology`` submodule is archived in Zenodo with the linked DOI:[@ZenodoLithologySourceCode] (note: doi not active for initial JOSS submission).
+Source code for ``Lithology`` and ``Litholayers`` is available as part of the [Landlab python package](https://github.com/landlab/landlab) in the [``Lithology`` submodule](https://github.com/landlab/landlab/tree/master/landlab/components/lithology). The ``Lithology`` submodule is documented using Docstrings, and the documentation can be found on the Landlab ReadTheDocs site. One page exists for the [Lithology component](https://landlab.readthedocs.io/en/master/landlab.components.lithology.html) and a second for the [LithoLayers component](https://landlab.readthedocs.io/en/master/landlab.components.litholayers.html). Unit and docstring tests provide 100% coverage of this submodule. [Pull Request # 674](https://github.com/landlab/landlab/pull/674) brought the ``Lithology`` submodule into the core Landlab source code. A tagged version of Landlab that includes the ``Lithology`` submodule is archived in Zenodo with the linked DOI:[@ZenodoLithologySourceCode] (note: doi not active for initial JOSS submission).
 
 The Landlab project maintains a separate repository containing tutorials that introduce core concepts and the use of individual submodules. In addition to the source code, a [Jupyter Notebook introducing the use of Lithology and Litholayers](https://github.com/landlab/tutorials/blob/next/lithology/lithology_and_litholayers.ipynb) is now part of the Landlab tutorials repository. This tutorial was brought into the repository with [Pull Request # 19](https://github.com/landlab/tutorials/pull/19). A tagged version of Landlab that includes this functionality has been archived with the following DOI: [@ZenodoLithologyNotebook] (note: doi not active for initial JOSS submission).
 
