+++ landlab/components/lithology/joss/papers.bib
@@ -54,25 +54,44 @@
 }
 
 @Article{ZenodoLithologySourceCode,
-    Abstractnote = {},
-    Author = {},
-    Date-Modified = {},
-    Doi = {},
-    Month = {},
-    Publisher = {Zenodo},
-    Title = {},
-    Year = {},
-    Bdsk-Url-1 = {}
+	author       = {Eric Hutton and
+	Dan Hobley and
+	Katy Barnhart and
+	Greg Tucker and
+	Sai Siddhartha Nudurupati and
+	Jordan Adams and
+	Nicole M Gasparini and
+	Charlie Shobe and
+	Ronda Strauch and
+	Jenny Knuth and
+	Nathan Lyons and
+	Rachel Glade and
+	Giuseppecipolla95 and
+	Amanda Manaster and
+	margauxmouchene and
+	Francis Rengers},
+	title        = {landlab/landlab},
+	month        = oct,
+	year         = 2018,
+	doi          = {10.5281/zenodo.1451936},
+  url          = {https://doi.org/10.5281/zenodo.1451936}
 }
 
 @Article{ZenodoLithologyNotebook,
-    Abstractnote = {},
-    Author = {},
-    Date-Modified = {},
-    Doi = {},
-    Month = {},
-    Publisher = {Zenodo},
-    Title = {},
-    Year = {},
-    Bdsk-Url-1 = {}
-    }
+  author       = {Jenny Knuth and
+                  Katy Barnhart and
+                  Eric Hutton and
+                  Greg Tucker and
+                  Giuseppecipolla95 and
+                  Sai Siddhartha Nudurupati and
+                  Dan Hobley and
+                  margauxmouchene and
+                  Jordan Adams and
+                  Nicole M Gasparini and
+                  Nathan Lyons},
+  title        = {landlab/tutorials v1.5.4-alpha},
+  month        = oct,
+  year         = 2018,
+  doi          = {10.5281/zenodo.1451948},
+  url          = {https://doi.org/10.5281/zenodo.1451948}
+}
