### 2018-10-08 715097c
updating bibliography (incl missing DOIS)
