+++ landlab/components/lithology/joss/papers.bib
@@ -24,74 +24,88 @@
     }
 
 @Article{Perne2017Steady,
-    title={Steady state, erosional continuity, and the topography of landscapes developed in layered rocks},
-    author={Perne, Matija and Covington, Matthew D and Thaler, Evan A and Myre, Joseph M},
-    journal={Earth Surface Dynamics},
-    volume={5},
-    number={1},
-    pages={85--100},
-    year={2017},
-    publisher={Copernicus GmbH}
+    AUTHOR = {Perne, M. and Covington, M. D. and Thaler, E. A. and Myre, J. M.},
+    TITLE = {Steady state, erosional continuity, and the topography of landscapes developed in layered rocks},
+    JOURNAL = {Earth Surface Dynamics},
+    VOLUME = {5},
+    YEAR = {2017},
+    NUMBER = {1},
+    PAGES = {85--100},
+    URL = {https://www.earth-surf-dynam.net/5/85/2017/},
+    DOI = {10.5194/esurf-5-85-2017}
     }
 
 @article{braun2013very,
-      title={A very efficient O (n), implicit and parallel method to solve the stream power equation governing fluvial incision and landscape evolution},
-      author={Braun, Jean and Willett, Sean D},
-      journal={Geomorphology},
-      volume={180},
-      pages={170--179},
-      year={2013},
-      publisher={Elsevier}
+    title = "A very efficient O(n), implicit and parallel method to solve the stream power equation governing fluvial incision and landscape evolution",
+    journal = "Geomorphology",
+    volume = "180-181",
+    pages = "170 - 179",
+    year = "2013",
+    issn = "0169-555X",
+    doi = "https://doi.org/10.1016/j.geomorph.2012.10.008",
+    url = "http://www.sciencedirect.com/science/article/pii/S0169555X12004618",
+    author = "Jean Braun and Sean D. Willett",
+    keywords = "Fluvial geomorphology, Numerical modeling, Stream power equation, Routing algorithm, Implicit scheme, Parallel algorithm"
     }
 
-@incollection{tucker2001channel,
-    title={The channel-hillslope integrated landscape development model (CHILD)},
-    author={Tucker, Gregory and Lancaster, Stephen and Gasparini, Nicole and Bras, Rafael},
-    booktitle={Landscape erosion and evolution modeling},
-    pages={349--388},
-    year={2001},
-    publisher={Springer}
-}
+@Inbook{tucker2001channel,
+    author="Tucker, Gregory
+    and Lancaster, Stephen
+    and Gasparini, Nicole
+    and Bras, Rafael",
+    editor="Harmon, Russell S.
+    and Doe, William W.",
+    title="The Channel-Hillslope Integrated Landscape Development Model (CHILD)",
+    bookTitle="Landscape Erosion and Evolution Modeling",
+    year="2001",
+    publisher="Springer US",
+    address="Boston, MA",
+    pages="349--388",
[…]
