### 2018-10-09 0a1cd7a
attributes and methods are not well rendered by RTFD. So I'm removing them.
