+++ landlab/components/lithology/litholayers.py
@@ -38,25 +38,6 @@ class LithoLayers(Lithology):
     are rock type IDs. The rock type IDs can be any type that is valid as a
     python dictionary key.
 
-    The following attributes and methods are the same as in the Lithology.
-
-    Attributes
-    ----------
-    z_top
-    z_bottom
-    thickness
-    dz
-    tracked_properties
-    properties
-
-    Methods
-    -------
-    add_rock_type
-    add_rock_property
-    update_rock_properties
-    add_layer
-    run_one_step
-
     """
     _name = 'LithoLayers'
 
