### 2018-10-09 landlab/landlab#807 @kbarnhart
OK. I'm deciding to keep the rendered images in.
