### 2018-10-09 a898c84
small text changes including removing the dois not yet ready statments
