+++ landlab/components/lithology/joss/paper.md
@@ -33,17 +33,73 @@ bibliography: papers.bib
 
 # Summary
 
-The surface of the Earth reflects the competing advection of rock by tectonic processes and the erosion of rock by wind, water, and ice. Rock properties influence erosion rates by changing the processes responsible for erosion and the rate at which rock is weathered, detached, and turned into mobile sediment. Variations in the rock properties over space and with depth reflect the legacy of sedimentary deposition and tectonic deformation. Long-term landscape evolution modeling experiments that include the impact of spatially and temporally variable rock characteristics can be used to identify the impact of rock strength patterns on other geologic observables such as topography, erosion rates, and detrital mineral records [e.g., @Forte2016Complexites; @Perne2017Steady]. Identifying these relationships allows for better interpretations of the geologic record.
+The surface of the Earth reflects the competing advection of rock by tectonic
+processes and the erosion of rock by wind, water, and ice. Rock properties
+influence erosion rates by changing the processes responsible for erosion and
+the rate at which rock is weathered, detached, and turned into mobile sediment.
+Variations in the rock properties over space and with depth reflect the legacy
+of sedimentary deposition and tectonic deformation. Long-term landscape
+evolution modeling experiments that include the impact of spatially and
+temporally variable rock characteristics can be used to identify the impact of
+rock strength patterns on other geologic observables such as topography, erosion
+rates, and detrital mineral records [e.g., @Forte2016Complexites;
+@Perne2017Steady]. Identifying these relationships allows for better
+interpretations of the geologic record.
 
-Landlab is an Open Source Python package that provides a framework for the development of 2D numerical models, typically in Earth surface dynamics [@Hobley2017Creative]. Landlab was designed as a  modular framework in which different process components can be mixed and matched to construct a model based on a user's needs. Prior work on spatially variable lithology in landscape evolution has been done using a modified version of the channel-hillslope integrated landscape development [CHILD, @tucker2001channel] model [e.g., @Forte2016Complexites] and the FastScape V5 model [@braun2013very; @Perne2017Steady]. To provide these capabilities within the Landlab framework, there is a need for a Landlab submodule that can treat spatial variations in rock materials.  
+Landlab is an Open Source Python package that provides a framework for the
+development of 2D numerical models, typically in Earth surface dynamics
+[@Hobley2017Creative]. Landlab was designed as a  modular framework in which
+different process components can be mixed and matched to construct a model based
+on a user's needs. Prior work on spatially variable lithology in landscape
+evolution has been done using a modified version of the channel-hillslope
+integrated landscape development [CHILD, @tucker2001channel] model [e.g.,
+@Forte2016Complexites] and the FastScape V5 model [@braun2013very;
+@Perne2017Steady]. To provide these capabilities within the Landlab framework,
+there is a need for a Landlab submodule that can treat spatial variations in
+rock materials.  
 
[…]
