### 2018-10-10 ac8ad81
add citation information to components
