### 2018-10-10 #979 @kbarnhart
@kthyng I'm almost done but have a final question related to DOIs.

Once I know how to deal with the DOIs correctly, then the only thing I have left to do (after the reviewers confirm there is nothing left to do) is make a Landlab release that is tagged v1.5.4.

The questions I have are in two broad categories:
1.  How to best deal with archiving when the majority of the submission is in the Landlab source code repository, but an additional part of it is in the Landlab tutorials repository.
2. How to best deal with reserve DOIs for repositories with Zenodo and Github linked by webhooks.

The work here is in two repositories (rather than one). This is because the Landlab source code is in one repo and the tutorials are in a second. The real *software* contribution is the source code, so I think the way to deal with this is to list the archive as the source cod Zenodo repository.

On the second issue: We have Github and Zenodo webhook-linked for landlab _and_ my research indicates that it is not  possible to make a pre-reserved DOI for a GitHub repo that is webhook-linked to Zenodo. I have thus made references in the text to the _concept DOIs_ on Zenodo for these two repositories. I think that this makes sense to link to the concept DOIs, but let me know if you think I should do otherwise.

Unless I make a special Zenodo archive just for this manuscript (that would not be listed with the rest of the Landlab Zenodo DOIs under the Landlab concept DOI) I don't think I can make a reserve DOI.

However, a DOI specific to Landlab 1.5.4 will be generated when I create the release. I can respond with that DOI when you ask for the archive DOI at the end of the review process.

Does it make sense to plan to proceed in that way?
