### 2018-10-11 #979 @kyleniemeyer
@kbarnhart regarding this question:

>However, a DOI specific to Landlab 1.5.4 will be generated when I create the release. I can respond with that DOI when you ask for the archive DOI at the end of the review process.
>
>Does it make sense to plan to proceed in that way?

Yes, we just need the DOI at the end of the process, once all changes have been made and the submission is accepted.
