### 2018-10-11 #979 @kyleniemeyer
@kbarnhart for your question #1, for JOSS you should archive the source code and provide that DOI. For the tutorials, I think you can do two things:
- Link to the location from the source code (e.g., in the README), and
- Archive the tutorials separately and reference that DOI in your paper
