### 2018-10-11 #979 @jedbrown
If the tutorials are considered a key component (and especially if they have been part of the review), then I would recommend @kyleniemeyer's second option: create a DOI and reference it in the paper.
