### 2018-10-11 #979 @kbarnhart
Thanks for the thoughts @kyleniemeyer and @jedbrown ! Based on your suggestions (and because the tutorials have been part of the review) this is what I think makes sense. Let me know if you think otherwise.

Once I verify that the reviewers are happy with the revisions I will.

1. Create a release of the tutorials repository and get the release specific zenodo DOI for it.
2. Update the manuscript to reflect
    a. that the tutorials repository DOI is the release-specific v1.5.4
    b. that the landlab source code DOI referenced in the text is the concept DOI but the archive link is the release specific DOI.
   c. Do appropriate merging and then make a release of the landlab source code at v1.5.4
   d. Provide the v1.5.4 release specific DOI here to be used as the archive.
