### 2018-10-11 #979 @kthyng
Thanks for your input @kyleniemeyer and @jedbrown!

@kbarnhart Sounds like you have a good plan. I think we just need to hear back from @speleophysics with the go ahead and then we'll be good to move forward.
