### 2018-10-15 #979 @kbarnhart
@speleophysics would it be OK if I moved ahead with making the release?
