### 2018-10-15 7db2583
Clast_tracker with DataRecord
- delete file CenteredSlope (experimental)
- implement Clast tracker using DataRecord structure
- major changes in clast tracker functions: proba_detach, burial
