+++ landlab/utils/CenteredSlope_for_CT2.py
@@ -1,383 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Mar 27 10:14:52 2018
-
-@author: margaux
-"""
-
-import numpy as np
-import gc
-from landlab.item_collection import ItemCollection
-from math import pi
-
-
-class ClastCollection(ItemCollection):
-    """
-    """
-    def __init__(self,
-                 grid,
-                 clast_x=[],
-                 clast_y=[],
-                 clast_elev=[],
-                 clast_radius=[]):
-        """
-        """
-
-        # Save a reference to the grid
-        self._grid = grid
-#
-#        # Store data
-#        self._clast__x = clast_x
-#        self._clast__y = clast_y
-#        self._clast__elev = clast_elev
-#        self._clast__radius = clast_radius
-#
-
-        # Determine reference (closest) node for each clast
-        self._clast__node = self._grid.find_nearest_node((self._clast__x, self._clast__y))
-        # Clast set size:
-        self._nb_of_clast = len(self._clast__node)
-
-        # Store the input information and others in a dictionary:
-        
-        clast_data = {'clast__x' : clast_x,
-                      'clast__y' : clast_y,
-                      'clast__elev' : clast_elev,
-                      'clast__node' : self._clast__node,
-                      'lambda_0' : np.zeros(self._nb_of_clast),
-                      'lambda_mean' : np.zeros(self._nb_of_clast),
-                      'slope__WE' : np.zeros(self._nb_of_clast),
-                      'slope__SN' : np.zeros(self._nb_of_clast),
-                      'slope__steepest_azimuth' : np.full(self._nb_of_clast, np.NaN),
-                      'slope__steepest_dip' : np.zeros(self._nb_of_clast),
-                      'distance__to_exit' : np.full(self._nb_of_clast, np.NaN),
-                      'target_node' : -np.ones(self._nb_of_clast, dtype=int),
-                      'change_x' : np.zeros(self._nb_of_clast),
-                      'change_y' : np.zeros(self._nb_of_clast),
-                      'hop_length' : np.zeros(self._nb_of_clast),
-                      'total_travelled_dist' : np.zeros(self._nb_of_clast)
-                      }
-
-        # Build ItemCollection containing clast data:
-        ItemCollection.__init__(self,               # necessary?
-                                self._grid,
-                                data=clast_data,
-                                grid_element='node',
-                                element_id=self._clast__node)
-
-
-    def _neighborhood(self, clast):
-        self._clast__node = self._grid.find_nearest_node((self._clast__x, self._clast__y)) # needs this to update after clast has moved
-        _node = self._clast__node[clast]
-        _grid = self._grid
-        ### Calculation of slopes: W to E and S to N, units=m/m
-        we_slope=(_grid.at_node['topographic__elevation'][west_node]-
-                  _grid.at_node['topographic__elevation'][east_node])/(2*rmg.dx)
-        
-        sn_slope=(self._grid.at_node['topographic__elevation'][south_node]-
-                  self._grid.at_node['topographic__elevation'][north_node])/(2*self._grid.dy)
-        
-        
-        ### COORDINATES OF CORNERS OF CELL
-        x_of_corners_at_node=np.zeros((self._grid.number_of_nodes,4))
[…]
