+++ landlab/item_collection/clast_collection_DR.py
@@ -4,14 +4,14 @@
 
 import numpy as np
 import gc
-from landlab.item_collection import ItemCollection
+from landlab.data_record import DataRecord
 from matplotlib.pyplot import figure, show, plot, xlabel, ylabel, title, legend
 
 from math import pi
 
 from scipy.stats import truncnorm
 
-class ClastCollection(ItemCollection):
+class ClastCollection(DataRecord):
     """
     ClastCollection is a datastructure inheriting from ItemCollection and
     including methods to move the clasts.
@@ -287,13 +287,19 @@ class ClastCollection(ItemCollection):
         # Determine reference (closest) node for each clast:
         _clast__node=[]
         _clast__node[:] = self._grid.find_nearest_node((clast_x[:], clast_y[:]))
-
+        # Clast collection size:
+        self._nb_of_clast = len(_clast__node)
+        # CHeck clast elev (not above node elev):
+        _clast__elev=clast_elev
+        for clast in range(self._nb_of_clast):
+            if _clast__elev[clast] > \
+            self._grid.at_node['topographic__elevation'][_clast__node[clast]]:
+                _clast__elev[clast] = \
+                self._grid.at_node['topographic__elevation'][
+                        _clast__node[clast]] # snaps clast to grid if was above
         # Determine current cell for each clast:
         _clast__cell=[]
         _clast__cell[:] = self._grid.cell_at_node[_clast__node]
-
-        # Clast collection size:
-        self._nb_of_clast = len(_clast__node)
 
         # Store the input information in a dictionary and create the other
         # fields that will be populated later:
@@ -326,34 +332,40 @@ class ClastCollection(ItemCollection):
         # total_travelled_dist = total distance that the clast has travelled
         #######################################################################
 
-        clast_data = {'clast__x' : clast_x,
-                      'clast__y' : clast_y,
-                      'clast__elev' : clast_elev,
-                      'clast__node' : _clast__node,
-                      'clast__initial_radius' : clast_radius,
-                      'clast__radius' : clast_radius,
-                      'lambda_0' : np.zeros(self._nb_of_clast),
-                      'lambda_mean' : np.zeros(self._nb_of_clast),
-                      'slope__WE' : np.zeros(self._nb_of_clast),
-                      'slope__SN' : np.zeros(self._nb_of_clast),
-                      'slope__steepest_azimuth' : np.full(
-                              self._nb_of_clast, np.NaN),
-                      'slope__steepest_dip' : np.zeros(self._nb_of_clast),
-                      'distance__to_exit' : np.full(self._nb_of_clast, np.NaN),
-                      'target_node' : -np.ones(self._nb_of_clast, dtype=int),
-                      'target_node_flag': -np.ones(
-                              self._nb_of_clast, dtype=int),
-                      'change_x' : np.zeros(self._nb_of_clast), #remove?
-                      'change_y' : np.zeros(self._nb_of_clast), #remove?
-                      'hop_length' : np.zeros(self._nb_of_clast),
-                      'total_travelled_dist' : np.zeros(self._nb_of_clast)}
+        clast_data = {'clast__x' : (['item_id', 'time'], clast_x.reshape((self._nb_of_clast,1))),
+                      'clast__y' : (['item_id', 'time'], clast_y.reshape((self._nb_of_clast,1))),
+                      'clast__elev' : (['item_id', 'time'], clast_elev.reshape((self._nb_of_clast,1))),
+                      'clast__node' : (['item_id'], np.array(_clast__node)),
+                      'clast__initial_radius' : (['item_id'], clast_radius),
+                      'clast__radius' : (['item_id', 'time'], clast_radius.reshape((self._nb_of_clast,1))),
[…]
