### 2018-10-15 0c309d3
Fix AppVeyor badge branch
AppVeyor badge was pointing to last built rather than to master.

This meant that because a recent feature branch build failed, the AppVeyor badge indicated a failure (even though the most recent `master` branch build was a success).
