### 2018-10-15 671be10
adding properties of closed nodes
