+++ landlab/components/flow_director/flow_direction_dinf.py
@@ -275,7 +275,7 @@ def flow_directions_dinf(grid, elevs="topographic__elevation", baselevel_nodes=N
 
     ##### Step 4, Initialize receiver and proportion arrays
     receivers = UNDEFINED_INDEX * np.ones((num_nodes, num_receivers), dtype=int)
-    #receiver_closed = UNDEFINED_INDEX * np.ones((num_nodes, num_receivers), dtype=int)
+    receiver_closed = UNDEFINED_INDEX * np.ones((num_nodes, num_receivers), dtype=int)
     proportions = np.zeros((num_nodes, num_receivers), dtype=float)
     receiver_links = UNDEFINED_INDEX * np.ones((num_nodes, num_receivers), dtype=int)
     slopes_to_receivers = np.zeros((num_nodes, num_receivers), dtype=float)
@@ -349,9 +349,10 @@ def flow_directions_dinf(grid, elevs="topographic__elevation", baselevel_nodes=N
     steepest_rg = np.empty_like(node_id, dtype=float)
     steepest_s = np.empty_like(node_id, dtype=float)
 
+    closed_triangle_neighbors = closed_nodes[triangle_neighbors_at_node]
     for n in node_id:
         steepest_rg[n] = rg[n, steepest_sort[n, -1]]
-        #receiver_closed[n] = closed_triangle_neighbors[n, :, steepest_sort[n, -1]]
+        receiver_closed[n] = closed_triangle_neighbors[n, :, steepest_sort[n, -1]]
         steepest_s[n] = s[n, steepest_sort[n, -1]]
         receivers[n, :] = triangle_neighbors_at_node[n, :, steepest_sort[n, -1]]
         receiver_links[n, :] = triangle_links_at_node[n, :, steepest_sort[n, -1]]
@@ -361,7 +362,6 @@ def flow_directions_dinf(grid, elevs="topographic__elevation", baselevel_nodes=N
 
     # construct the baseline for proportions
     rg_baseline = np.array([0., 1., 1., 2., 2., 3., 3., 4]) * np.pi / 2.
-    # rg_baseline = np.array([0., 0.5, 1., 1.5, 2., 2.5, 3., 3.5])*np.pi/4.
 
     # calculate alpha1 and alpha 2
     alpha2 = (steepest_rg - rg_baseline[steepest_triangle]) * af[steepest_triangle]
@@ -370,6 +370,9 @@ def flow_directions_dinf(grid, elevs="topographic__elevation", baselevel_nodes=N
     # calculate proportions from alpha
     proportions[:, 0] = (alpha1) / (alpha1 + alpha2)
     proportions[:, 1] = (alpha2) / (alpha1 + alpha2)
+
+    # where proportions == 0, set reciever  to -1
+    receivers[proportions==0] = -1
 
     ### END OF THE Tarboton algorithm, start of work to make this code mesh
     # with other landlab flow directing algorithms.
@@ -384,30 +387,30 @@ def flow_directions_dinf(grid, elevs="topographic__elevation", baselevel_nodes=N
     drains_to_self[steepest_s < 0] = True
 
     # if both receiver nodes are closed, drain to self
-    #drains_to_two_closed = receiver_closed.sum(axis=1) == num_receivers
-    #drains_to_self[drains_to_two_closed] = True
+    drains_to_two_closed = receiver_closed.sum(axis=1) == num_receivers
+    drains_to_self[drains_to_two_closed] = True
 
     # if drains to one closed receiver, check that the open receiver actually
     # gets flow. If so, route all to the open receiver. If the receiver getting
     # all the flow is closed, then drain to self.
-    #all_flow_to_closed = np.sum(receiver_closed * proportions, axis=1) == 1
[…]
