### 2018-10-16 af69bc5
more tests, and algo wip
