+++ landlab/components/flow_director/tests/test_mfd.py
@@ -1,8 +1,10 @@
 
 import pytest
 import numpy as np
+from numpy.testing import assert_array_equal
 
 from landlab import RasterModelGrid
+from landlab.components import FlowDirectorMFD
 from landlab.components.flow_director import flow_direction_mfd
 
 
@@ -25,3 +27,21 @@ def test_bad_argument_mfd():
             link_slope,
             partition_method="foo",
         )
+
+def test_mfd_on_flat_terrain():
+    mg = RasterModelGrid((5, 4), spacing=(1, 1))
+    mg.add_zeros('node','topographic__elevation')
+
+    fd = FlowDirectorMFD(mg)
+    fd.run_one_step()
+
+    node_ids = np.arange(mg.number_of_nodes)
+    true_recievers = -1 * np.ones(fd.receivers.shape)
+    true_recievers[:, 0] = node_ids
+
+    true_proportions = np.zeros(fd.proportions.shape)
+    true_proportions[:, 0] = 1
+
+    assert_array_equal(fd.receivers, true_recievers)
+    assert_array_equal(np.round(fd.proportions, decimals=6),
+                       np.round(true_proportions, decimals=6))
