+++ landlab/components/flow_director/flow_director_dinf.py
@@ -77,18 +77,18 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [ 1, -1],
            [ 2, -1],
            [ 3, -1],
-           [ 0,  1],
-           [ 1,  0],
+           [ 4, -1],
+           [-1,  0],
            [ 5,  1],
-           [ 6,  2],
+           [ 7, -1],
            [ 8, -1],
            [ 5, -1],
-           [ 6,  5],
-           [10,  6],
+           [-1,  5],
+           [11, -1],
            [12, -1],
-           [ 9, -1],
-           [10,  9],
-           [-1, 10]])
+           [13, -1],
+           [14, -1],
+           [15, -1]])
 
     It also stores the proportions of flow going to each receiver, the link on
     which the flow moves in at node arrays, and the slope of each link.
@@ -101,15 +101,15 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [ 1.        ,  0.        ],
            [ 0.        ,  1.        ],
            [ 0.59033447,  0.40966553],
-           [ 0.74866817,  0.25133183],
+           [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 0.        ,  1.        ],
-           [ 0.31191652,  0.68808348],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
-           [ 0.31191652,  0.68808348],
-           [ 0.        ,  1.        ]])
+           [ 1.        ,  0.        ],
+           [ 1.        ,  0.        ],
+           [ 1.        ,  0.        ]])
     >>> mg.at_node['flow__link_to_receiver_node']
     array([[-1, -1],
            [-1, -1],
@@ -119,36 +119,36 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [ 4, 24],
            [ 8, 26],
            [ 9, 28],
-           [-1, -1],
-           [11, 30],
+           [14, 31],
+           [11, 33],
            [12, 32],
            [16, 34],
-           [-1, -1],
-           [18, 36],
+           [21, 37],
+           [18, 39],
            [19, 38],
            [20, 40]])
     >>> mg.at_node['topographic__steepest_slope'] # doctest: +NORMALIZE_WHITESPACE
-    array([[-1.        ,  0.        ],
-           [ 1.        , -0.        ],
-           [ 3.        ,  1.41421356],
-           [ 5.        ,  2.82842712],
-           [ 1.        ,  0.        ],
-           [ 1.        ,  1.41421356],
-           [ 3.        ,  2.82842712],
-           [ 5.        ,  4.24264069],
-           [ 3.        ,  1.41421356],
-           [ 3.        ,  2.82842712],
-           [ 3.        ,  4.24264069],
-           [ 5.        ,  5.65685425],
-           [ 5.        ,  2.82842712],
-           [ 5.        ,  4.24264069],
-           [ 5.        ,  5.65685425],
-           [ 5.        ,  7.07106781]])
+    array([[ -1.00000000e+00,  -1.41421356e+00],
+           [  1.00000000e+00,  -7.12763635e+02],
+           [  3.00000000e+00,   1.41421356e+00],
+           [  5.00000000e+00,   2.82842712e+00],
+           [  1.00900000e+03,   7.12763635e+02],
+           [  1.00000000e+00,   1.41421356e+00],
+           [  3.00000000e+00,   2.82842712e+00],
+           [  1.00400000e+03,   7.10642315e+02],
+           [  1.00400000e+03,   7.12056529e+02],
[…]
