### 2018-10-16 89435bb
adding tests. black
