+++ landlab/components/flow_director/tests/test_mfd.py
@@ -28,9 +28,10 @@ def test_bad_argument_mfd():
             partition_method="foo",
         )
 
+
 def test_mfd_on_flat_terrain():
     mg = RasterModelGrid((5, 4), spacing=(1, 1))
-    mg.add_zeros('node','topographic__elevation')
+    mg.add_zeros("node", "topographic__elevation")
 
     fd = FlowDirectorMFD(mg)
     fd.run_one_step()
@@ -43,5 +44,60 @@ def test_mfd_on_flat_terrain():
     true_proportions[:, 0] = 1
 
     assert_array_equal(fd.receivers, true_recievers)
-    assert_array_equal(np.round(fd.proportions, decimals=6),
-                       np.round(true_proportions, decimals=6))
+    assert_array_equal(
+        np.round(fd.proportions, decimals=6), np.round(true_proportions, decimals=6)
+    )
+
+
+def test_mfd_flat_closed_lower():
+    mg = RasterModelGrid((5, 4), spacing=(1, 1))
+    z = mg.add_zeros("node", "topographic__elevation")
+    z[mg.core_nodes] += 1
+    mg.set_closed_boundaries_at_grid_edges(
+        bottom_is_closed=True,
+        left_is_closed=True,
+        right_is_closed=True,
+        top_is_closed=True,
+    )
+
+    fd = FlowDirectorMFD(mg)
+    fd.run_one_step()
+
+    node_ids = np.arange(mg.number_of_nodes)
+    true_recievers = -1 * np.ones(fd.receivers.shape)
+    true_recievers[:, 0] = node_ids
+
+    true_proportions = np.zeros(fd.proportions.shape)
+    true_proportions[:, 0] = 1
+
+    assert_array_equal(fd.receivers, true_recievers)
+    assert_array_equal(
+        np.round(fd.proportions, decimals=6), np.round(true_proportions, decimals=6)
+    )
+
+
+def test_mfd_flat_closed_upper():
+    mg = RasterModelGrid((5, 4), spacing=(1, 1))
+    z = mg.add_zeros("node", "topographic__elevation")
+    z[mg.core_nodes] -= 1
+    mg.set_closed_boundaries_at_grid_edges(
+        bottom_is_closed=True,
+        left_is_closed=True,
+        right_is_closed=True,
+        top_is_closed=True,
+    )
+
+    fd = FlowDirectorMFD(mg)
+    fd.run_one_step()
+
+    node_ids = np.arange(mg.number_of_nodes)
+    true_recievers = -1 * np.ones(fd.receivers.shape)
+    true_recievers[:, 0] = node_ids
+
+    true_proportions = np.zeros(fd.proportions.shape)
+    true_proportions[:, 0] = 1
+
+    assert_array_equal(fd.receivers, true_recievers)
+    assert_array_equal(
+        np.round(fd.proportions, decimals=6), np.round(true_proportions, decimals=6)
+    )
