### 2018-10-16 7a4818a
Homogenize input format, improve doc, improve unit tests and coverage
