### 2018-10-16 15af211
improvements, tests, black
