### 2018-10-16 43c48d2
finishing getting the order correct
I now ensure that if flow is going to -1, this occurs in the second column. THis requires moving around some arrays, and swapping values in the docstrings.
