+++ landlab/components/flow_director/flow_director_dinf.py
@@ -78,12 +78,12 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [ 2, -1],
            [ 3, -1],
            [ 4, -1],
-           [-1,  0],
+           [ 0,  -1],
            [ 5,  1],
            [ 7, -1],
            [ 8, -1],
            [ 5, -1],
-           [-1,  5],
+           [ 5, -1],
            [11, -1],
            [12, -1],
            [13, -1],
@@ -99,12 +99,12 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
-           [ 0.        ,  1.        ],
+           [ 1.        ,  0.        ],
            [ 0.59033447,  0.40966553],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
-           [ 0.        ,  1.        ],
+           [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
@@ -116,12 +116,12 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [-1, -1],
            [-1, -1],
            [ 3, 25],
-           [ 4, 24],
+           [24,  4],
            [ 8, 26],
            [ 9, 28],
            [14, 31],
            [11, 33],
-           [12, 32],
+           [32, 12],
            [16, 34],
            [21, 37],
            [18, 39],
@@ -133,12 +133,12 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [  3.00000000e+00,   1.41421356e+00],
            [  5.00000000e+00,   2.82842712e+00],
            [  1.00900000e+03,   7.12763635e+02],
-           [  1.00000000e+00,   1.41421356e+00],
+           [  1.41421356e+00,   1.00000000e+00],
            [  3.00000000e+00,   2.82842712e+00],
            [  1.00400000e+03,   7.10642315e+02],
            [  1.00400000e+03,   7.12056529e+02],
            [  3.00000000e+00,   0.00000000e+00],
-           [  3.00000000e+00,   4.24264069e+00],
+           [  4.24264069e+00,   3.00000000e+00],
            [  1.00100000e+03,   7.09935208e+02],
            [ -0.00000000e+00,   7.09935208e+02],
            [  1.00400000e+03,   7.07813888e+02],
@@ -162,12 +162,12 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [ 2, -1],
            [ 3, -1],
            [ 4, -1],
-           [-1,  0],
+           [ 0, -1],
            [ 5,  1],
            [ 7, -1],
            [ 8, -1],
            [ 5, -1],
-           [-1,  5],
+           [ 5, -1],
            [11, -1],
            [12, -1],
            [13, -1],
@@ -179,12 +179,12 @@ class FlowDirectorDINF(_FlowDirectorToMany):
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
-           [ 0.        ,  1.        ],
+           [ 1.        ,  0.        ],
            [ 0.59033447,  0.40966553],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
-           [ 0.        ,  1.        ],
+           [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
            [ 1.        ,  0.        ],
