### 2018-10-16 1a9e079
few more swapped elements
