+++ landlab/components/flow_director/tests/test_mfd.py
@@ -1,10 +1,10 @@
 
 import pytest
 import numpy as np
-from numpy.testing import assert_array_equal
+from numpy.testing import assert_array_equal, assert_array_almost_equal
 
 from landlab import RasterModelGrid
-from landlab.components import FlowDirectorMFD
+from landlab.components import FlowDirectorMFD, FlowAccumulator
 from landlab.components.flow_director import flow_direction_mfd
 
 
@@ -101,3 +101,108 @@ def test_mfd_flat_closed_upper():
     assert_array_equal(
         np.round(fd.proportions, decimals=6), np.round(true_proportions, decimals=6)
     )
+
+
+def test_MFD_SW_slope():
+    mg = RasterModelGrid(10, 10, spacing=(1, 1))
+    z = mg.add_field("topographic__elevation", mg.node_y + mg.node_x, at="node")
+    fa = FlowAccumulator(mg, flow_director="FlowDirectorMFD")
+    fa.run_one_step()
+
+    # this one should flow half to the west and half to the south
+    w_links = mg.adjacent_nodes_at_node[:, 2]
+    s_links = mg.adjacent_nodes_at_node[:, 3]
+    node_ids = np.arange(mg.number_of_nodes)
+    true_recievers = -1 * np.ones(fa.flow_director.receivers.shape)
+    true_recievers[mg.core_nodes, 2] = w_links[mg.core_nodes]
+    true_recievers[mg.core_nodes, 3] = s_links[mg.core_nodes]
+    true_recievers[mg.boundary_nodes, 0] = node_ids[mg.boundary_nodes]
+
+    true_proportions = np.zeros(fa.flow_director.proportions.shape)
+    true_proportions[mg.boundary_nodes, 0] = 1
+    true_proportions[mg.core_nodes, 2:] = 0.5
+
+    assert_array_equal(true_recievers, fa.flow_director.receivers)
+    assert_array_equal(true_proportions, fa.flow_director.proportions)
+
+
+def test_MFD_SW_slope_w_diags():
+    mg = RasterModelGrid(10, 10, spacing=(1, 1))
+    z = mg.add_field("topographic__elevation", mg.node_y + mg.node_x, at="node")
+    fa = FlowAccumulator(mg, flow_director="FlowDirectorMFD", diagonals=True)
+    fa.run_one_step()
+
+    # this one should part to south west, part to south, and part to west
+    sw_diags = mg.diagonal_adjacent_nodes_at_node[:, 2]
+    w_links = mg.adjacent_nodes_at_node[:, 2]
+    s_links = mg.adjacent_nodes_at_node[:, 3]
+    node_ids = np.arange(mg.number_of_nodes)
+    true_recievers = -1 * np.ones(fa.flow_director.receivers.shape)
+    true_recievers[mg.core_nodes, 2] = w_links[mg.core_nodes]
+    true_recievers[mg.core_nodes, 3] = s_links[mg.core_nodes]
+    true_recievers[mg.core_nodes, 6] = sw_diags[mg.core_nodes]
+    true_recievers[mg.boundary_nodes, 0] = node_ids[mg.boundary_nodes]
+
+    true_proportions = np.zeros(fa.flow_director.proportions.shape)
+    true_proportions[mg.boundary_nodes, 0] = 1
+
+    total_sum_of_slopes = 1. + 1. + (2. / 2. ** 0.5)
+
+    true_proportions[mg.core_nodes, 2] = 1.0 / total_sum_of_slopes
+    true_proportions[mg.core_nodes, 3] = 1.0 / total_sum_of_slopes
[…]
