### 2018-10-17 33a4151
update to flow accum to n
turns out that this had not been designed to handle a grid in which every single node flows to itself. Now it can handle it.
