+++ landlab/components/flow_accum/tests/test_flow_accumulator.py
@@ -1028,17 +1028,18 @@ def test_hex_mfd():
 
 def test_flat_grids_all_directors():
     for fd in [
+        "FlowDirectorMFD",
         "FlowDirectorSteepest",
         "FlowDirectorD8",
         "FlowDirectorDINF",
-        "FlowDirectorMFD",
+
     ]:
-        mg = RasterModelGrid(10, 20, dx=1)
+        mg = RasterModelGrid(10, 10)
         z = mg.add_zeros("topographic__elevation", at="node")
         fa = FlowAccumulator(mg, flow_director=fd)
-        try:
-            fa.run_one_step()
-            print(fd + " success")
-        except IndexError:
-            print(fd + " failed")
+        fa.run_one_step()
+
+        true_da = np.zeros(mg.size('node'))
+        true_da[mg.core_nodes] = 1.0
+        assert_array_equal(true_da, fa.drainage_area)
         del mg, z, fa
