### 2018-10-17 #979 @kthyng
I just read through @speleophysics previous comments and I think everything has been satisfied. I feel comfortable moving forward.

@kbarnhart I'm setting your paper as accepted now! Can you create an archive (on Zenodo, figshare, or other) and post the archive DOI here?
