+++ landlab/components/lithology/joss/papers.bib
@@ -69,43 +69,43 @@
 
 @Article{ZenodoLithologySourceCode,
     author       = {Eric Hutton and
-    Dan Hobley and
-    Katy Barnhart and
-    Greg Tucker and
-    Sai Siddhartha Nudurupati and
-    Jordan Adams and
-    Nicole M Gasparini and
-    Charlie Shobe and
-    Ronda Strauch and
-    Jenny Knuth and
-    Nathan Lyons and
-    Rachel Glade and
-    Giuseppecipolla95 and
-    Amanda Manaster and
-    margauxmouchene and
-    Francis Rengers},
+                    Dan Hobley and
+                    Katy Barnhart and
+                    Greg Tucker and
+                    Sai Siddhartha Nudurupati and
+                    Jordan Adams and
+                    Nicole M Gasparini and
+                    Charlie Shobe and
+                    Ronda Strauch and
+                    Jenny Knuth and
+                    Nathan Lyons and
+                    Rachel Glade and
+                    Giuseppecipolla95 and
+                    Amanda Manaster and
+                    margauxmouchene and
+                    Francis Rengers},
     title        = {landlab/landlab},
-    month        = oct,
+    month        = Oct,
     year         = 2018,
     doi          = {10.5281/zenodo.1451936},
     url          = {https://doi.org/10.5281/zenodo.1451936}
     }
 
 @Article{ZenodoLithologyNotebook,
-    author       = {Jenny Knuth and
-                    Katy Barnhart and
-                    Eric Hutton and
-                    Greg Tucker and
-                    Giuseppecipolla95 and
-                    Sai Siddhartha Nudurupati and
-                    Dan Hobley and
-                    margauxmouchene and
-                    Jordan Adams and
-                    Nicole M Gasparini and
-                    Nathan Lyons},
-    title        = {landlab/tutorials v1.5.4-alpha},
-    month        = oct,
-    year         = 2018,
-    doi          = {10.5281/zenodo.1451948},
-    url          = {https://doi.org/10.5281/zenodo.1451948}
+      author       = {Katy Barnhart and
+                      Jenny Knuth and
+                      Eric Hutton and
+                      Greg Tucker and
+                      Giuseppecipolla95 and
+                      Sai Siddhartha Nudurupati and
+                      Dan Hobley and
+                      margauxmouchene and
+                      Jordan Adams and
+                      Nicole M Gasparini and
+                      Nathan Lyons},
+      title        = {landlab/tutorials v1.5.4},
+      month        = Oct,
+      year         = 2018,
+      doi          = {10.5281/zenodo.1464296},
+      url          = {https://doi.org/10.5281/zenodo.1464296}
     }
