### 2018-10-17 9f4386f
paper updates to correct DOI language
