+++ landlab/components/lithology/joss/paper.md
@@ -85,8 +85,9 @@ component](https://landlab.readthedocs.io/en/release/landlab.components.litholay
 Unit and docstring tests provide 100% coverage of this submodule. [Pull Request #
 674](https://github.com/landlab/landlab/pull/674) brought the ``Lithology``
 submodule into the core Landlab source code. The first release version of
-Landlab that includes the ``Lithology`` submodule is tagged as v1.5.4. It is
-archived in Zenodo with the linked DOI: [@ZenodoLithologySourceCode].
+Landlab that includes the ``Lithology`` submodule is tagged as v1.5.4. The
+concept DOI for Landlab is archived in Zenodo with the linked DOI: [@ZenodoLithologySourceCode]
+and the archive for this manuscript points to the Zenodo archive of v1.5.4.
 
 The Landlab project maintains a separate repository containing tutorials that
 introduce core concepts and the use of individual submodules. In addition to the
@@ -95,8 +96,8 @@ Litholayers](https://nbviewer.jupyter.org/github/landlab/tutorials/blob/release/
 is now part of the Landlab tutorials repository. This tutorial was brought into
 the repository with [Pull Request #
 19](https://github.com/landlab/tutorials/pull/19). The first release version of
-the Landlab tutorials that includes this notebooks is tagged as v1.5.4. It is
-archived with the following DOI: [@ZenodoLithologyNotebook].
+the Landlab tutorials that includes this notebooks is tagged as v1.5.4 and is
+archived in Zenodo with the linked DOI: [@ZenodoLithologyNotebook].
 
 # Acknowledgements
 
