### 2018-10-17 landlab/landlab#814 @kbarnhart
Final DOI updates needed in JOSS manuscript
The [JOSS review](https://github.com/openjournals/joss-reviews/issues/979#issuecomment-430447632) identified the right way to do DOI linking and this needs to be implemented.
