### 2018-10-17 #979 @kthyng
Hey @arfon we have an accepted paper here!!

Thanks to @speleophysics and @tristan-salles for their reviews, and congrats to @kbarnhart!
