### 2018-10-17 #979 @speleophysics
Yes. I am/was ok with acceptance. I've been out of contact doing field work
the last week.
-------------------------
Matt Covington
Associate Professor
Department of Geosciences
340 N. Campus Dr.
216 Gearhart Hall
University of Arkansas
Fayetteville, AR 72701

webpage - http://www.speleophysics.com/
-------------------------

On Tue, Oct 16, 2018 at 10:29 PM Kristen Thyng <notifications@github.com>
wrote:

> I just read through @speleophysics <https://github.com/speleophysics>
> previous comments and I think everything has been satisfied. I feel
> comfortable moving forward.
>
> @kbarnhart <https://github.com/kbarnhart> I'm setting your paper as
> accepted now! Can you create an archive (on Zenodo, figshare, or other) and
> post the archive DOI here?
>
