### 2018-10-17 #979 @arfon
:wave: @kthyng - there are a couple of steps I need to take before we should close this issue. I'll do that now...
